+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 6.0.0-dev
 - first assertion failure in a test method now ends the method's execution
 - raised minimal version of PHP to 8.3
+- improved output of results formatter TextDox in console
 
 Version 5.1.0
 - added assertions assertNoException and assertMatchesFile